refactor: Remove 'go-' prefix from go-unit-test folder names
- Rename go-httpmock -> httpmock
- Rename go-mockery -> mockery
- Rename go-testcontainers -> testcontainers
- Update import paths in affected Go files
- Update references in CLAUDE.md
